pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-otel"
version = "0.1.0"
description = "Seamless OpenTelemetry integration for Django"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Sivakamy Lakshminarayanan", email = "letscodewithshiva@gmail.com"}
]
keywords = ["django", "opentelemetry", "tracing", "monitoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
]

dependencies = [
  "Django>=3.2",
  "opentelemetry-api>=1.20",
  "opentelemetry-sdk>=1.22",
  "opentelemetry-exporter-otlp>=1.20",
  "azure-monitor-opentelemetry>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/yourusername/django-otel"
Documentation = "https://github.com/yourusername/django-otel#readme"
Repository = "https://github.com/yourusername/django-otel"
Issues = "https://github.com/yourusername/django-otel/issues"

[tool.setuptools.packages.find]
where = ["src"]
